+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [1.1.0] - Not Yet Released
 
+### Added
+- New selector in performances and Web Vitals analytics to differentiate frontend and backend pages.
+
 ## [1.0.3] - 2021-12-16
 
 ### Changed